Add an contrived example of a waiter leak
If we track the running things for our waiter in module scope then we
run the risk of one test timing out and then poisoning all subsequent
tests.

You can see that tests B and C work fine if you filter to run just the tests
in that module but if test A has already run then the bad state is stored
in module scope and they give misleading failures.