Demonstration that `registerWaiter` can also leak
And poison subsequent tests if it is used naively. The waiter doesn't get
torn down so `wait` in a subsequent test still waits for it (see that the
`console.log`s keep logging after the tests are finished).

I had incorrectly assumed that all waiters would be explictly removed
between test runs but that seems not to be the case.